API

As stated in the Home page, we consider the nonlinear optimization problem in the following format:

\[\begin{align*} \min \quad & f(x) \\ & c_L \leq c(x) \leq c_U \\ & \ell \leq x \leq u. \end{align*}\]

To develop an optimization algorithm, we are usually worried not only with $f(x)$ and $c(x)$, but also with their derivatives. Namely,

$\nabla f(x)$, the gradient of $f$ at the point $x$;

$\nabla^2 f(x)$, the Hessian of $f$ at the point $x$;

$J(x) = \nabla c(x)$, the Jacobian of $c$ at the point $x$;

$\nabla^2 f(x) + \sum_{i=1}^m \lambda_i \nabla^2 c_i(x)$, the Hessian of the Lagrangian function at the point $(x,\lambda)$.

There are many ways to access some of these values, so here is a little reference guide.

Reference guide

The following naming should be easy enough to follow. If not, click on the link and go to the description.

! means inplace;

_coord means coordinate format;

prod means matrix-vector product;

_op means operator (as in LinearOperators.jl).

Feel free to open an issue to suggest other methods that should apply to all NLPModels instances.

Function	NLPModels function

$f(x)$	obj, objgrad, objgrad!, objcons, objcons!

$\nabla f(x)$	grad, grad!, objgrad, objgrad!

$\nabla^2 f(x)$	hess, hess_op, hess_op!, hess_coord, hess_coord, hess_structure, hess_structure!, hprod, hprod!

$c(x)$	cons, cons!, objcons, objcons!

$J(x)$	jac, jac_op, jac_op!, jac_coord, jac_coord!, jac_structure, jprod, jprod!, jtprod, jtprod!

$\nabla^2 L(x,y)$	hess, hess_op, hess_coord, hess_coord!, hess_structure, hess_structure!, hprod, hprod!

API for NLSModels

For the Nonlinear Least Squares models, $f(x) = \Vert F(x)\Vert^2$, and these models have additional function to access the residual value and its derivatives. Namely,

$J_F(x) = \nabla F(x)$

$\nabla^2 F_i(x)$

Function	function

$F(x)$	residual, residual!

$J_F(x)$	jac_residual, jac_coord_residual, jac_coord_residual!, jac_structure_residual, jprod_residual, jprod_residual!, jtprod_residual, jtprod_residual!, jac_op_residual, jac_op_residual!

$\nabla^2 F_i(x)$	hess_residual, hess_coord_residual, hess_coord_residual!, hess_structure_residual, hess_structure_residual!, jth_hess_residual, hprod_residual, hprod_residual!, hess_op_residual, hess_op_residual!

AbstractNLPModel functions

NLPModels.obj — Function

f = obj(nlp, x)

Evaluate $f(x)$, the objective function of nlp at x.

NLPModels.grad — Function

g = grad(nlp, x)

Evaluate $∇f(x)$, the gradient of the objective function at x.

NLPModels.grad! — Function

g = grad!(nlp, x, g)

Evaluate $∇f(x)$, the gradient of the objective function at x in place.

NLPModels.objgrad — Function

f, g = objgrad(nlp, x)

Evaluate $f(x)$ and $∇f(x)$ at x.

NLPModels.objgrad! — Function

f, g = objgrad!(nlp, x, g)

Evaluate $f(x)$ and $∇f(x)$ at x. g is overwritten with the value of $∇f(x)$.

NLPModels.cons — Function

c = cons(nlp, x)

Evaluate $c(x)$, the constraints at x.

NLPModels.cons! — Function

c = cons!(nlp, x, c)

Evaluate $c(x)$, the constraints at x in place.

NLPModels.objcons — Function

f, c = objcons(nlp, x)

Evaluate $f(x)$ and $c(x)$ at x.

NLPModels.objcons! — Function

f = objcons!(nlp, x, c)

Evaluate $f(x)$ and $c(x)$ at x. c is overwritten with the value of $c(x)$.

NLPModels.jac_coord — Function

vals = jac_coord(nlp, x)

Evaluate $∇c(x)$, the constraint's Jacobian at x in sparse coordinate format.

NLPModels.jac_coord! — Function

vals = jac_coord!(nlp, x, vals)

Evaluate $∇c(x)$, the constraint's Jacobian at x in sparse coordinate format, rewriting vals.

NLPModels.jac_structure — Function

(rows,cols) = jac_structure(nlp)

Return the structure of the constraint's Jacobian in sparse coordinate format.

NLPModels.jac_structure! — Function

jac_structure!(nlp, rows, cols)

Return the structure of the constraint's Jacobian in sparse coordinate format in place.

NLPModels.jac — Function

Jx = jac(nlp, x)

Evaluate $∇c(x)$, the constraint's Jacobian at x as a sparse matrix.

NLPModels.jac_op — Function

J = jac_op(nlp, x)

Return the Jacobian at x as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v.

NLPModels.jac_op! — Function

J = jac_op!(nlp, x, Jv, Jtv)

Return the Jacobian at x as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v. The values Jv and Jtv are used as preallocated storage for the operations.

J = jac_op!(nlp, rows, cols, vals, Jv, Jtv)

Return the Jacobian given by (rows, cols, vals) as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v. The values Jv and Jtv are used as preallocated storage for the operations.

J = jac_op!(nlp, x, rows, cols, Jv, Jtv)

Return the Jacobian at x as a linear operator. The resulting object may be used as if it were a matrix, e.g., J * v or J' * v. (rows, cols) should be the sparsity structure of the Jacobian. The values Jv and Jtv are used as preallocated storage for the operations.

NLPModels.jprod — Function

Jv = jprod(nlp, x, v)

Evaluate $∇c(x)v$, the Jacobian-vector product at x.

NLPModels.jprod! — Function

Jv = jprod!(nlp, x, v, Jv)

Evaluate $∇c(x)v$, the Jacobian-vector product at x in place.

Jv = jprod!(nlp, rows, cols, vals, v, Jv)

Evaluate $∇c(x)v$, the Jacobian-vector product, where the Jacobian is given by (rows, cols, vals) in triplet format.

Jv = jprod!(nlp, x, rows, cols, v, Jv)

Evaluate $∇c(x)v$, the Jacobian-vector product at x in place. (rows, cols) should be the Jacobian structure in triplet format.

NLPModels.jtprod — Function

Jtv = jtprod(nlp, x, v, Jtv)

Evaluate $∇c(x)^Tv$, the transposed-Jacobian-vector product at x.

NLPModels.jtprod! — Function

Jtv = jtprod!(nlp, x, v, Jtv)

Evaluate $∇c(x)^Tv$, the transposed-Jacobian-vector product at x in place.

Jtv = jtprod!(nlp, rows, cols, vals, v, Jtv)

Evaluate $∇c(x)^Tv$, the transposed-Jacobian-vector product, where the Jacobian is given by (rows, cols, vals) in triplet format.

Jtv = jtprod!(nlp, x, rows, cols, v, Jtv)

Evaluate $∇c(x)^Tv$, the transposed-Jacobian-vector product at x in place. (rows, cols) should be the Jacobian structure in triplet format.

NLPModels.hess_coord — Function

vals = hess_coord(nlp, x; obj_weight=1.0)

Evaluate the objective Hessian at x in sparse coordinate format, with objective function scaled by obj_weight, i.e.,

\[σ ∇²f(x),\]

with σ = obj_weight . Only the lower triangle is returned.

vals = hess_coord(nlp, x, y; obj_weight=1.0)

Evaluate the Lagrangian Hessian at (x,y) in sparse coordinate format, with objective function scaled by obj_weight, i.e.,

\[∇²L(x,y) = σ ∇²f(x) + ∑ᵢ yᵢ ∇²cᵢ(x),\]

with σ = obj_weight . Only the lower triangle is returned.

NLPModels.hess_coord! — Function

vals = hess_coord!(nlp, x, vals; obj_weight=1.0)

Evaluate the objective Hessian at x in sparse coordinate format, with objective function scaled by obj_weight, i.e.,

\[σ ∇²f(x),\]

with σ = obj_weight , rewriting vals. Only the lower triangle is returned.

vals = hess_coord!(nlp, x, y, vals; obj_weight=1.0)

Evaluate the Lagrangian Hessian at (x,y) in sparse coordinate format, with objective function scaled by obj_weight, i.e.,

\[∇²L(x,y) = σ ∇²f(x) + ∑ᵢ yᵢ ∇²cᵢ(x),\]

with σ = obj_weight , rewriting vals. Only the lower triangle is returned.

NLPModels.hess_structure — Function

(rows,cols) = hess_structure(nlp)

Return the structure of the Lagrangian Hessian in sparse coordinate format.

NLPModels.hess_structure! — Function

hess_structure!(nlp, rows, cols)

Return the structure of the Lagrangian Hessian in sparse coordinate format in place.

NLPModels.hess — Function

Hx = hess(nlp, x; obj_weight=1.0)

Evaluate the objective Hessian at x as a sparse matrix, with objective function scaled by obj_weight, i.e.,

\[σ ∇²f(x),\]

with σ = obj_weight . Only the lower triangle is returned.

Hx = hess(nlp, x, y; obj_weight=1.0)

Evaluate the Lagrangian Hessian at (x,y) as a sparse matrix, with objective function scaled by obj_weight, i.e.,

\[∇²L(x,y) = σ ∇²f(x) + ∑ᵢ yᵢ ∇²cᵢ(x),\]

with σ = obj_weight . Only the lower triangle is returned.

NLPModels.hess_op — Function

H = hess_op(nlp, x; obj_weight=1.0)

Return the objective Hessian at x with objective function scaled by obj_weight as a linear operator. The resulting object may be used as if it were a matrix, e.g., H * v. The linear operator H represents

\[σ ∇²f(x),\]

with σ = obj_weight .

H = hess_op(nlp, x, y; obj_weight=1.0)

Return the Lagrangian Hessian at (x,y) with objective function scaled by obj_weight as a linear operator. The resulting object may be used as if it were a matrix, e.g., H * v. The linear operator H represents

\[∇²L(x,y) = σ ∇²f(x) + ∑ᵢ yᵢ ∇²cᵢ(x),\]

with σ = obj_weight .

NLPModels.hess_op! — Function

H = hess_op!(nlp, x, Hv; obj_weight=1.0)

Return the objective Hessian at x with objective function scaled by obj_weight as a linear operator, and storing the result on Hv. The resulting object may be used as if it were a matrix, e.g., w = H * v. The vector Hv is used as preallocated storage for the operation. The linear operator H represents

\[σ ∇²f(x),\]

with σ = obj_weight .

H = hess_op!(nlp, rows, cols, vals, Hv)

Return the Hessian given by (rows, cols, vals) as a linear operator, and storing the result on Hv. The resulting object may be used as if it were a matrix, e.g., w = H * v. The vector Hv is used as preallocated storage for the operation. The linear operator H represents

\[σ ∇²f(x),\]

with σ = obj_weight .

H = hess_op!(nlp, x, rows, cols, Hv; obj_weight=1.0)

Return the objective Hessian at x with objective function scaled by obj_weight as a linear operator, and storing the result on Hv. The resulting object may be used as if it were a matrix, e.g., w = H * v. (rows, cols) should be the sparsity structure of the Hessian. The vector Hv is used as preallocated storage for the operation. The linear operator H represents

\[σ ∇²f(x),\]

with σ = obj_weight .

H = hess_op!(nlp, x, y, Hv; obj_weight=1.0)

Return the Lagrangian Hessian at (x,y) with objective function scaled by obj_weight as a linear operator, and storing the result on Hv. The resulting object may be used as if it were a matrix, e.g., w = H * v. The vector Hv is used as preallocated storage for the operation. The linear operator H represents

\[∇²L(x,y) = σ ∇²f(x) + ∑ᵢ yᵢ ∇²cᵢ(x),\]

with σ = obj_weight .

H = hess_op!(nlp, x, y, rows, cols, Hv; obj_weight=1.0)

Return the Lagrangian Hessian at (x,y) with objective function scaled by obj_weight as a linear operator, and storing the result on Hv. The resulting object may be used as if it were a matrix, e.g., w = H * v. (rows, cols) should be the sparsity structure of the Hessian. The vector Hv is used as preallocated storage for the operation. The linear operator H represents

\[σ ∇²f(x),\]

with σ = obj_weight .

NLPModels.hprod — Function

Hv = hprod(nlp, x, v; obj_weight=1.0)

Evaluate the product of the objective Hessian at x with the vector v, with objective function scaled by obj_weight, where the objective Hessian is

\[σ ∇²f(x),\]

with σ = obj_weight .

Hv = hprod(nlp, x, y, v; obj_weight=1.0)

Evaluate the product of the Lagrangian Hessian at (x,y) with the vector v, with objective function scaled by obj_weight, where the Lagrangian Hessian is

\[∇²L(x,y) = σ ∇²f(x) + ∑ᵢ yᵢ ∇²cᵢ(x),\]

with σ = obj_weight .

NLPModels.hprod! — Function

Hv = hprod!(nlp, x, v, Hv; obj_weight=1.0)

Evaluate the product of the objective Hessian at x with the vector v in place, with objective function scaled by obj_weight, where the objective Hessian is

\[σ ∇²f(x),\]

with σ = obj_weight .

Hv = hprod!(nlp, rows, cols, vals, v, Hv)

Evaluate the product of the objective or Lagrangian Hessian given by (rows, cols, vals) in triplet format with the vector v in place. Only one triangle of the Hessian should be given.

Hv = hprod!(nlp, x, rows, cols, v, Hv)

Evaluate the product of the objective Hessian at x with the vector v in place, where the objective Hessian is

\[σ ∇²f(x),\]

with σ = obj_weight . (rows, cols) should be the Hessian structure in triplet format.

Hv = hprod!(nlp, x, y, v, Hv; obj_weight=1.0)

Evaluate the product of the Lagrangian Hessian at (x,y) with the vector v in place, with objective function scaled by obj_weight, where the Lagrangian Hessian is

\[∇²L(x,y) = σ ∇²f(x) + ∑ᵢ yᵢ ∇²cᵢ(x),\]

with σ = obj_weight .

Hv = hprod!(nlp, x, y, rows, cols, v, Hv)

Evaluate the product of the Lagrangian Hessian at (x,y) with the vector v in place, where the Lagrangian Hessian is

\[∇²L(x,y) = σ ∇²f(x) + ∑ᵢ yᵢ ∇²cᵢ(x),\]

with σ = obj_weight . (rows, cols) should be the Hessian structure in triplet format.

LinearOperators.reset! — Function

reset!(counters)

Reset evaluation counters

reset!(nlp)

Reset evaluation count in nlp

Base.print — Function

print(io, meta)

Prints meta information - x0, nvar, ncon, etc.

AbstractNLSModel

NLPModels.residual — Function

Fx = residual(nls, x)

Computes F(x), the residual at x.

NLPModels.residual! — Function

Fx = residual!(nls, x, Fx)

Computes F(x), the residual at x.

NLPModels.jac_residual — Function

Jx = jac_residual(nls, x)

Computes J(x), the Jacobian of the residual at x.

NLPModels.jac_coord_residual — Function

(rows,cols,vals) = jac_coord_residual(nls, x)

Computes the Jacobian of the residual at x in sparse coordinate format.

NLPModels.jac_coord_residual! — Function

vals = jac_coord_residual!(nls, x, vals)

Computes the Jacobian of the residual at x in sparse coordinate format, rewriting vals. rows and cols are not rewritten.

NLPModels.jac_structure_residual — Function

(rows,cols) = jac_structure_residual(nls)

Returns the structure of the constraint's Jacobian in sparse coordinate format.

NLPModels.jac_structure_residual! — Function

(rows,cols) = jac_structure_residual!(nls, rows, cols)

Returns the structure of the constraint's Jacobian in sparse coordinate format in place.

NLPModels.jprod_residual — Function

Jv = jprod_residual(nls, x, v)

Computes the product of the Jacobian of the residual at x and a vector, i.e., J(x)*v.

NLPModels.jprod_residual! — Function

Jv = jprod_residual!(nls, x, v, Jv)

Computes the product of the Jacobian of the residual at x and a vector, i.e., J(x)*v, storing it in Jv.

Jv = jprod_residual!(nls, rows, cols, vals, v, Jv)

Computes the product of the Jacobian of the residual given by (rows, cols, vals) and a vector, i.e., J(x)*v, storing it in Jv.

Jv = jprod_residual!(nls, x, rows, cols, v, Jv)

Computes the product of the Jacobian of the residual at x and a vector, i.e., J(x)*v, storing it in Jv. The structure of the Jacobian is given by (rows, cols).

NLPModels.jtprod_residual — Function

Jtv = jtprod_residual(nls, x, v)

Computes the product of the transpose of the Jacobian of the residual at x and a vector, i.e., J(x)'*v.

NLPModels.jtprod_residual! — Function

Jtv = jtprod_residual!(nls, x, v, Jtv)

Computes the product of the transpose of the Jacobian of the residual at x and a vector, i.e., J(x)'*v, storing it in Jtv.

Jtv = jtprod_residual!(nls, rows, cols, vals, v, Jtv)

Computes the product of the transpose of the Jacobian of the residual given by (rows, cols, vals) and a vector, i.e., J(x)'*v, storing it in Jv.

Jtv = jtprod_residual!(nls, x, rows, cols, v, Jtv)

Computes the product of the transpose Jacobian of the residual at x and a vector, i.e., J(x)'*v, storing it in Jv. The structure of the Jacobian is given by (rows, cols).

NLPModels.jac_op_residual — Function

Jx = jac_op_residual(nls, x)

Computes J(x), the Jacobian of the residual at x, in linear operator form.

NLPModels.jac_op_residual! — Function

Jx = jac_op_residual!(nls, x, Jv, Jtv)

Computes J(x), the Jacobian of the residual at x, in linear operator form. The vectors Jv and Jtv are used as preallocated storage for the operations.

Jx = jac_op_residual!(nls, rows, cols, vals, Jv, Jtv)

Computes J(x), the Jacobian of the residual given by (rows, cols, vals), in linear operator form. The vectors Jv and Jtv are used as preallocated storage for the operations.

Jx = jac_op_residual!(nls, x, rows, cols, Jv, Jtv)

Computes J(x), the Jacobian of the residual at x, in linear operator form. The vectors Jv and Jtv are used as preallocated storage for the operations. The structure of the Jacobian should be given by (rows, cols).

NLPModels.hess_residual — Function

H = hess_residual(nls, x, v)

Computes the linear combination of the Hessians of the residuals at x with coefficients v.

NLPModels.hess_coord_residual — Function

vals = hess_coord_residual(nls, x, v)

Computes the linear combination of the Hessians of the residuals at x with coefficients v in sparse coordinate format.

NLPModels.hess_coord_residual! — Function

vals = hess_coord_residual!(nls, x, v, vals)

Computes the linear combination of the Hessians of the residuals at x with coefficients v in sparse coordinate format, rewriting vals.

NLPModels.hess_structure_residual — Function

(rows,cols) = hess_structure_residual(nls)

Returns the structure of the residual Hessian.

NLPModels.hess_structure_residual! — Function

hess_structure_residual!(nls, rows, cols)

Returns the structure of the residual Hessian in place.

NLPModels.jth_hess_residual — Function

Hj = jth_hess_residual(nls, x, j)

Computes the Hessian of the j-th residual at x.

NLPModels.hprod_residual — Function

Hiv = hprod_residual(nls, x, i, v)

Computes the product of the Hessian of the i-th residual at x, times the vector v.

NLPModels.hprod_residual! — Function

Hiv = hprod_residual!(nls, x, i, v, Hiv)

Computes the product of the Hessian of the i-th residual at x, times the vector v, and stores it in vector Hiv.

NLPModels.hess_op_residual — Function

Hop = hess_op_residual(nls, x, i)

Computes the Hessian of the i-th residual at x, in linear operator form.

NLPModels.hess_op_residual! — Function

Hop = hess_op_residual!(nls, x, i, Hiv)

Computes the Hessian of the i-th residual at x, in linear operator form. The vector Hiv is used as preallocated storage for the operation.

Derivative Checker

NLPModels.gradient_check — Function

Check the first derivatives of the objective at x against centered finite differences.

This function returns a dictionary indexed by components of the gradient for which the relative error exceeds rtol.

NLPModels.jacobian_check — Function

Check the first derivatives of the constraints at x against centered finite differences.

This function returns a dictionary indexed by (j, i) tuples such that the relative error in the i-th partial derivative of the j-th constraint exceeds rtol.

NLPModels.hessian_check — Function

Check the second derivatives of the objective and each constraints at x against centered finite differences. This check does not rely on exactness of the first derivatives, only on objective and constraint values.

The sgn arguments refers to the formulation of the Lagrangian in the problem. It should have a positive value if the Lagrangian is formulated as

L(x,y) = f(x) + ∑ yⱼ cⱼ(x)

e.g., as in JuMPNLPModels, and a negative value if the Lagrangian is formulated as

L(x,y) = f(x) - ∑ yⱼ cⱼ(x)

e.g., as in AmplModels. Only the sign of sgn is important.

This function returns a dictionary indexed by functions. The 0-th function is the objective while the k-th function (for k > 0) is the k-th constraint. The values of the dictionary are dictionaries indexed by tuples (i, j) such that the relative error in the second derivative ∂²fₖ/∂xᵢ∂xⱼ exceeds rtol.

NLPModels.hessian_check_from_grad — Function

Check the second derivatives of the objective and each constraints at x against centered finite differences. This check assumes exactness of the first derivatives.

The sgn arguments refers to the formulation of the Lagrangian in the problem. It should have a positive value if the Lagrangian is formulated as

L(x,y) = f(x) + ∑ yⱼ cⱼ(x)

e.g., as in JuMPNLPModels, and a negative value if the Lagrangian is formulated as

L(x,y) = f(x) - ∑ yⱼ cⱼ(x)

e.g., as in AmplModels. Only the sign of sgn is important.

This function returns a dictionary indexed by functions. The 0-th function is the objective while the k-th function (for k > 0) is the k-th constraint. The values of the dictionary are dictionaries indexed by tuples (i, j) such that the relative error in the second derivative ∂²fₖ/∂xᵢ∂xⱼ exceeds rtol.

Internal

NLPModels.coo_prod! — Function

coo_prod!(rows, cols, vals, v, Av)

Compute the product of a matrix A given by (rows, cols, vals) and the vector v. The result is stored in Av, which should have length equals to the number of rows of A.

NLPModels.coo_sym_prod! — Function

coo_sym_prod!(rows, cols, vals, v, Av)

Compute the product of a symmetric matrix A given by (rows, cols, vals) and the vector v. The result is stored in Av, which should have length equals to the number of rows of A. Only one triangle of A should be passed.

NLPModels.increment! — Function

increment!(nlp, s)

Increment counter s of problem nlp.

NLPModels.decrement! — Function

decrement!(nlp, s)

Decrement counter s of problem nlp.