Models

Models

The following general models are implemented in this package:

ADNLPModel

Derived Models

SlackModel

LBFGSModel

LSR1Model

In addition, the following nonlinear least squares models are implemented in this package:

ADNLSModel

FeasibilityResidual

LLSModel

SlackNLSModel

FeasibilityFormNLS

There are other external models implemented. In particular,

AmplModel

CUTEstModel

MathProgNLPModel and MathProgNLSModel using JuMP/MPB.

There are currently two models implemented in this package, besides the external ones.

ADNLPModel

NLPModels.ADNLPModel — Type.

ADNLPModel is an AbstractNLPModel using ForwardDiff to compute the derivatives. In this interface, the objective function $f$ and an initial estimate are required. If there are constraints, the function $c:ℝⁿ → ℝᵐ$ and the vectors $c_L$ and $c_U$ also need to be passed. Bounds on the variables and an inital estimate to the Lagrangian multipliers can also be provided.

ADNLPModel(f, x0; lvar = [-∞,…,-∞], uvar = [∞,…,∞], y0 = zeros,
  c = NotImplemented, lcon = [-∞,…,-∞], ucon = [∞,…,∞], name = "Generic")

f - The objective function $f$. Should be callable;

x0 :: AbstractVector - The initial point of the problem;

lvar :: AbstractVector - $ℓ$, the lower bound of the variables;

uvar :: AbstractVector - $u$, the upper bound of the variables;

c - The constraints function $c$. Should be callable;

y0 :: AbstractVector - The initial value of the Lagrangian estimates;

lcon :: AbstractVector - $c_L$, the lower bounds of the constraints function;

ucon :: AbstractVector - $c_U$, the upper bounds of the constraints function;

name :: String - A name for the model.

The functions follow the same restrictions of ForwardDiff functions, summarised here:

The function can only be composed of generic Julia functions;

The function must accept only one argument;

The function's argument must accept a subtype of AbstractVector;

The function should be type-stable.

For contrained problems, the function $c$ is required, and it must return an array even when m = 1, and $c_L$ and $c_U$ should be passed, otherwise the problem is ill-formed. For equality constraints, the corresponding index of $c_L$ and $c_U$ should be the same.

Example

using NLPModels
f(x) = sum(x.^4)
x = [1.0; 0.5; 0.25; 0.125]
nlp = ADNLPModel(f, x)
grad(nlp, x)

4-element Array{Float64,1}:
 4.0
 0.5
 0.0625
 0.0078125

Derived Models

The following models are created from any given model, making some modification to that model.

SlackModel

NLPModels.SlackModel — Type.

A model whose only inequality constraints are bounds.

Given a model, this type represents a second model in which slack variables are introduced so as to convert linear and nonlinear inequality constraints to equality constraints and bounds. More precisely, if the original model has the form

\[\begin{align*} \min_x \quad & f(x)\\ \mathrm{s.t.} \quad & c_L ≤ c(x) ≤ c_U,\\ & ℓ ≤ x ≤ u, \end{align*}\]

the new model appears to the user as

\[\begin{align*} \min_X \quad & f(X)\\ \mathrm{s.t.} \quad & g(X) = 0,\\ & L ≤ X ≤ U. \end{align*}\]

The unknowns $X = (x, s)$ contain the original variables and slack variables $s$. The latter are such that the new model has the general form

\[\begin{align*} \min_x \quad & f(x)\\ \mathrm{s.t.} \quad & c(x) - s = 0,\\ & c_L ≤ s ≤ c_U,\\ & ℓ ≤ x ≤ u. \end{align*}\]

although no slack variables are introduced for equality constraints.

The slack variables are implicitly ordered as [s(low), s(upp), s(rng)], where low, upp and rng represent the indices of the constraints of the form $c_L ≤ c(x) < ∞$, $-∞ < c(x) ≤ c_U$ and $c_L ≤ c(x) ≤ c_U$, respectively.

Example

using NLPModels
f(x) = x[1]^2 + 4x[2]^2
c(x) = [x[1]*x[2] - 1]
x = [2.0; 2.0]
nlp = ADNLPModel(f, x, c=c, lcon=[0.0])
nlp_slack = SlackModel(nlp)
nlp_slack.meta.lvar

3-element Array{Float64,1}:
 -Inf
 -Inf
    0.0

LBFGSModel

NLPModels.LBFGSModel — Type.

Construct a LBFGSModel from another type of model.

LSR1Model

NLPModels.LSR1Model — Type.

Construct a LSR1Model from another type of nlp.

ADNLSModel

NLPModels.ADNLSModel — Type.

ADNLSModel is an Nonlinear Least Squares model using ForwardDiff to compute the derivatives.

ADNLSModel(F, x0, m; lvar = [-∞,…,-∞], uvar = [∞,…,∞], y0 = zeros,
  c = NotImplemented, lcon = [-∞,…,-∞], ucon = [∞,…,∞], name = "Generic")

F - The residual function $F$. Should be callable;

x0 :: AbstractVector - The initial point of the problem;

m :: Int - The dimension of $F(x)$, i.e., the number of

equations in the nonlinear system.

The other parameters are as in ADNLPModel.

using NLPModels
F(x) = [x[1] - 1; 10*(x[2] - x[1]^2)]
nlp = ADNLSModel(F, [-1.2; 1.0], 2)
residual(nlp, nlp.meta.x0)

2-element Array{Float64,1}:
 -2.2
 -4.3999999999999995

FeasibilityResidual

NLPModels.FeasibilityResidual — Type.

A feasibility residual model is created from a NLPModel of the form

min f(x)
s.t cℓ ≤ c(x) ≤ cu
    bℓ ≤   x  ≤ bu

by creating slack variables s and defining F(x,s) = c(x) - s. The resulting NLS problem is

min ¹/₂‖c(x) - s‖²
    bℓ ≤ x ≤ bu
    cℓ ≤ s ≤ bu

This is done using SlackModel first, and then defining the NLS. Notice that if bℓᵢ = buᵢ, no slack variable is created.

LLSModel

NLPModels.LLSModel — Type.

nls = LLSModel(A, b; lvar, uvar, C, lcon, ucon)

Creates a Linear Least Squares model ½‖Ax - b‖² with optional bounds lvar ≦ x ≦ y and optional linear constraints lcon ≦ Cx ≦ ucon.

SlackNLSModel

NLPModels.SlackNLSModel — Type.

Like SlackModel, this model converts inequalities into equalities and bounds.

FeasibilityFormNLS

NLPModels.FeasibilityFormNLS — Type.

Converts a nonlinear least-squares problem with residual F(x) to a nonlinear optimization problem with constraints F(x) = r and objective ¹/₂‖r‖². In other words, converts

min ¹/₂‖F(x)‖²
s.t  cₗ ≤ c(x) ≤ cᵤ
      ℓ ≤   x  ≤ u

to

min ¹/₂‖r‖²
s.t   F(x) - r = 0
     cₗ ≤ c(x) ≤ cᵤ
      ℓ ≤   x  ≤ u

If you rather have the first problem, the nls model already works as an NLPModel of that format.